requirements.txt
# ConvNeXt Earthquake Precursor Detection
# Requirements

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.62.0

# Configuration
PyYAML>=6.0
omegaconf>=2.1.0

# Logging
tensorboard>=2.8.0
wandb>=0.12.0  # optional

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# GradCAM visualization
grad-cam>=1.4.0
captum>=0.5.0

# Optional: faster data loading
albumentations>=1.1.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
